merge gavans-branch into main (#9)
* changed most footer icons to lucide icons and added lightmode/darkmode versions

* updated some colors and some styles

fixed my dum css syntax

* added logo, fixed shadows

Ive added the logo svg and fixed the dark mode shadows on the donate box, ive also changed the donate box to the primary color (red) for both themes. ive started work on replacing the <ImageCarousel> component with EmblaCarousel.

* more work on image carousel

Did some more work on the image carousel, i still need to style the buttons, add a autoplay progress, add scroll support, and also i need to make it work with the yaml config thing

* im not adding an autoplay bar

it will just stop autoplaying when the mouse enters focus or the buttons are interacted with

* button

* making a new sidebar

replacing the "jump to section" thing in the about page
also added scroll wheel gestures to the image carousel